# api/openapi.yaml
openapi: 3.0.3
info:
  title: transitforge API
  version: 0.5.0
  description: |
    Multi-node edge/proxy infrastructure manager.
    Authentication: Bearer token plus HMAC-SHA256 over
    `timestamp + method + path + sha256(body)` using the bearer token as key.
    Headers: `Authorization: Bearer <token>`, `X-TransitForge-Timestamp`, `X-TransitForge-Signature`.
    Roles: operator (mutate+read), readonly (read), agent (mutate+read), administrator (human UI; mutate+read+users).
    Human users: GET/POST /api/v1/users* (administrator, or self for TOTP/locale/password).
    Canonical contract: docs/ARCHITECTURE.md. This file must stay synchronized with it.
servers:
  - url: https://127.0.0.1:8443
paths:
  /healthz:
    get:
      summary: Liveness
      responses:
        "200":
          description: process alive
  /readyz:
    get:
      summary: Readiness (DB reachable, engine initialized)
      responses:
        "200":
          description: ready
        "503":
          description: not ready
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: exposition format
        "501":
          description: metrics not enabled
  /api/v1/nodes:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: list nodes
    post:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Node"
      responses:
        "201":
          description: created
  /api/v1/nodes/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: node
  /api/v1/nodes/{id}/desired-state:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: desired state (disabled mappings omitted)
  /api/v1/nodes/{id}/actual-state:
    get:
      summary: Read last stored actual-state
      description: Returns stored ActualState plus AgentStatus (TransportState is inside actual). Does not expose secrets. Empty object if the agent has not reported.
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: last reported actual-state and agent status (empty if none)
    post:
      summary: Agent reports actual-state
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "202":
          description: recorded
  /api/v1/nodes/{id}/plan:
    get:
      summary: Read-only reconcile plan
      description: |
        Reads desired-state and stored actual-state, runs reconcile.Diff, returns the plan.
        Does not mutate anything and does not create an apply audit event.
        Distinct from POST .../apply {"dry_run": true}, which is an explicit audited dry-run.
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: plan text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanView"
  /api/v1/nodes/{id}/apply:
    post:
      summary: Explicit apply (audited)
      description: |
        Operator-triggered apply. dry_run=true records apply-dry-run and returns the plan.
        When LiveApply is disabled, dry_run=false also returns a dry-run plan and audits apply-dry-run.
        Live host mutation is agent-driven.
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRequest"
      responses:
        "200":
          description: plan or apply result
        "409":
          description: CONFLICT
  /api/v1/nodes/{id}/failback:
    post:
      summary: Failback
      description: |
        Explicit human-triggered Failback: SSH_PRIMARY → WG_PRIMARY.
        The controller stores the intent and returns 202. The agent consumes it on the next cycle
        if WireGuard is healthy. Internal persistence may use action=fail_forward.
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                backend: {type: string}
                backend_id: {type: string}
                action: {type: string, description: implementation default fail_forward}
      responses:
        "202":
          description: Failback intent accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailbackIntent"
  /api/v1/backends:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: list
    post:
      security: [{bearerAuth: []}]
      responses:
        "201":
          description: created
  /api/v1/backends/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: backend
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: updated
  /api/v1/mappings:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: list (including disabled catalog entries)
    post:
      security: [{bearerAuth: []}]
      responses:
        "201":
          description: created (enabled defaults true)
  /api/v1/mappings/{id}:
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MappingPatch"
      responses:
        "200":
          description: mapping updated
    delete:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "204":
          description: deleted
  /api/v1/tunnels:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: list
    post:
      security: [{bearerAuth: []}]
      responses:
        "201":
          description: created
  /api/v1/tunnels/{id}/status:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: status
  /api/v1/sni-routes:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: list
    post:
      security: [{bearerAuth: []}]
      responses:
        "201":
          description: created
  /api/v1/sni-routes/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: route
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: updated
  /api/v1/events:
    get:
      summary: Audit events
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: node
          schema: {type: string}
          description: node id or name
        - in: query
          name: backend
          schema: {type: string}
          description: backend id or name
        - in: query
          name: since
          schema: {type: string}
          description: YYYY-MM-DD or RFC3339 (inclusive start). No from alias.
        - in: query
          name: until
          schema: {type: string}
          description: YYYY-MM-DD or RFC3339 (inclusive end of day for dates). No to alias.
        - in: query
          name: action
          schema: {type: string}
          description: exact audit action (create, update, apply-dry-run, failback, …)
      responses:
        "200":
          description: recent audit events
  /api/v1/whoami:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: caller name and role
  /api/v1/tokens:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: token metadata (administrator or operator; hashes omitted)
    post:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name: {type: string}
                role: {type: string, enum: [operator, readonly, agent]}
      responses:
        "201":
          description: minted token (plaintext returned once)
        "403":
          description: caller is not administrator or operator
  /api/v1/users:
    get:
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: human users (administrator; secrets omitted)
        "403":
          description: caller is not administrator
    post:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: {type: string}
                display_name: {type: string}
                password: {type: string}
                role: {type: string, enum: [administrator, operator, readonly]}
                locale: {type: string, enum: [en, ru]}
      responses:
        "201":
          description: created human user
  /api/v1/users/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: human user (administrator or self)
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: updated human user
  /api/v1/users/{id}/totp/begin:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: otpauth URL and secret (once)
  /api/v1/users/{id}/totp/confirm:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: TOTP enabled; recovery codes returned once
  /api/v1/users/{id}/totp/disable:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: TOTP disabled
  /api/v1/users/{id}/recovery-codes:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        "200":
          description: new recovery codes returned once
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Node:
      type: object
      required: [name]
      properties:
        name: {type: string}
        public_ip: {type: string}
        labels:
          type: object
          additionalProperties: {type: string}
    ApplyRequest:
      type: object
      properties:
        dry_run: {type: boolean}
    PlanView:
      type: object
      properties:
        plan: {type: string}
    MappingPatch:
      type: object
      properties:
        node_id: {type: string}
        backend_id: {type: string}
        backend: {type: string}
        protocol: {type: string, enum: [TCP, UDP]}
        public_port: {type: integer}
        backend_port: {type: integer}
        enabled:
          type: boolean
          description: |
            true: mapping participates in desired-state.
            false: mapping remains in inventory and is omitted from desired-state;
            the next plan DELETE's it if it still exists on the node.
    FailbackIntent:
      type: object
      properties:
        id: {type: string}
        node_id: {type: string}
        backend_id: {type: string}
        action: {type: string}
        created_at: {type: string, format: date-time}